Now that I've proved the basic concept, I have refactored the DynamicSupplierBuilder into a more manageable file and class structure and started work on providing default suppliers as well as handling primitive data type fields.
