+++ util/src/main/java/com/robertboothby/djenni/util/lambda/introspectable/IntrospectableSupplier.java
@@ -1,17 +1,16 @@
-package com.robertboothby.djenni.experimental;
+package com.robertboothby.djenni.util.lambda.introspectable;
 
 import java.io.Serializable;
 import java.lang.invoke.SerializedLambda;
 import java.lang.reflect.Method;
 import java.util.Arrays;
-import java.util.function.Function;
 import java.util.function.Supplier;
 
 /**
  * An introspectable Lambda function for a getter (or similar method).
  */
 @FunctionalInterface
-public interface IntrospectableFunction<T,R> extends Function<T,R>, Serializable {
+public interface IntrospectableSupplier<T> extends Supplier<T>, Serializable {
 
     default Class getImplementingClass() {
         try {
